pyproject.toml
[tool.poetry]
name = "tamingllms"
version = "0.1.0"
description = "accompanying package to \"Taming Language Models\" book"
authors = ["Tharsis T. P. Souza"]
license = "Attribution-NonCommercial-ShareAlike 4.0 International"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
jupyter-book = "^1.0.3"
pandas = "^2.2.3"
openai = "^1.54.4"
ipykernel = "^6.29.5"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
tiktoken = ">=0.5,<0.8"
litellm = "^1.52.9"
pydata-sphinx-theme = "^0.16.0"
sphinx-multitoc-numbering = "^0.1.3"
sphinxext-rediraffe = "^0.2.7"
sphinx-tojupyter = "^0.3.0"
sphinx-togglebutton = "^0.3.2"
insipid-sphinx-theme = "^0.4.2"
sphinx-press-theme = "^0.9.1"
transformers = "^4.46.3"
nbconvert = "^7.16.4"
evaluate = "^0.4.3"
absl-py = "^2.1.0"
rouge-score = "^0.1.2"
matplotlib = "^3.9.2"
outlines = "^0.1.7"
datasets = "^3.1.0"
text-generation = "^0.7.0"
ollama = "^0.4.3"
langsmith = "0.2.0"
sphinxcontrib-mermaid = "^1.0.0"
seaborn = "^0.13.2"
ipywidgets = "^8.1.5"
mistralai = "^1.2.5"
llm-guard = "^0.3.15"
pygments = "^2.18.0"
llama-cpp-python = "^0.3.5"
torch = "^2.5.1"
gguf = "^0.13.0"
accelerate = "^1.2.1"
markitdown = "^0.0.1a3"
docling = "^2.14.0"
python-levenshtein = "^0.26.1"
sphinx-math-dollar = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
